+++ components/frontend/src/header_footer/buttons/ResetSettingsButton.test.jsx
@@ -33,10 +33,12 @@ it("has no accessibility violations", async () => {
 
 it("resets the settings", async () => {
     history.push(
-        "?date_interval=2&date_order=ascending&hidden_columns=comment&hide_ignored_entities=metric_uuid&hidden_tags=tag&" +
-            "metrics_to_hide=no_action_required&nr_dates=2&show_issue_creation_date=true&show_issue_summary=true&" +
-            "show_issue_update_date=true&show_issue_due_date=true&show_issue_release=true&show_issue_sprint=true&" +
-            "sort_column=status&sort_direction=descending&expanded=tab:0&hidden_cards=tags",
+        "?date_interval=2&date_order=ascending&entity_sort_column=metric_uuid:status&" +
+            "entity_sort_direction=metric_uuid:descending&hidden_columns=comment&" +
+            "hide_ignored_entities=metric_uuid&hidden_tags=tag&metrics_to_hide=no_action_required&nr_dates=2&" +
+            "show_issue_creation_date=true&show_issue_summary=true&show_issue_update_date=true&" +
+            "show_issue_due_date=true&show_issue_release=true&show_issue_sprint=true&sort_column=status&" +
+            "sort_direction=descending&expanded=tab:0&hidden_cards=tags",
     )
     const settings = createTestableSettings()
     const handleDateChange = vi.fn()